feat : add syste modules entity data
