+++ app/GraphQL/Ecosystem/Mutations/Filesystem/AttachmentMutation.php
@@ -4,32 +4,28 @@ declare(strict_types=1);
 
 namespace App\GraphQL\Ecosystem\Mutations\Filesystem;
 
-use Kanvas\CustomFields\DataTransferObject\CustomFieldInput;
-use Kanvas\CustomFields\Repositories\CustomFieldsRepository;
 use Kanvas\Filesystem\Actions\AttachFilesystemAction;
 use Kanvas\Filesystem\DataTransferObject\FilesystemAttachInput;
 use Kanvas\Filesystem\Models\Filesystem;
 use Kanvas\Filesystem\Models\FilesystemEntities;
+use Kanvas\SystemModules\DataTransferObject\SystemModuleEntityInput;
+use Kanvas\SystemModules\Repositories\SystemModulesRepository;
 
 class AttachmentMutation
 {
     /**
      * Assign a filesystem to a entity.
-     *
-     * @param mixed $rootValue
-     * @param array $request
-     * @return string
      */
     public function attachFile(mixed $rootValue, array $request): string
     {
         $filesystemAttachmentInput = FilesystemAttachInput::viaRequest($request['input']);
-        $customFieldInput = new CustomFieldInput(
+        $entityInput = new SystemModuleEntityInput(
             'filesystem',
             $filesystemAttachmentInput->systemModuleUuid,
             $filesystemAttachmentInput->entityId
         );
 
-        $entity = CustomFieldsRepository::getEntityFromInput($customFieldInput, auth()->user());
+        $entity = SystemModulesRepository::getEntityFromInput($entityInput, auth()->user());
 
         $attachFile = new AttachFilesystemAction(
             Filesystem::getByUuid($filesystemAttachmentInput->filesystemUuid),
@@ -43,10 +39,6 @@ class AttachmentMutation
 
     /**
      * deAttach a file from filesystem
-     *
-     * @param mixed $rootValue
-     * @param array $request
-     * @return bool
      */
     public function deAttachFile(mixed $rootValue, array $request): bool
     {
