+++ src/Kanvas/CustomFields/DataTransferObject/CustomFieldInput.php
@@ -4,28 +4,11 @@ declare(strict_types=1);
 
 namespace Kanvas\CustomFields\DataTransferObject;
 
-use Spatie\LaravelData\Data;
+use Kanvas\SystemModules\DataTransferObject\SystemModuleEntityInput;
 
 /**
  * CustomFieldInput class.
  */
-class CustomFieldInput extends Data
+class CustomFieldInput extends SystemModuleEntityInput
 {
-    public function __construct(
-        public string $name,
-        public string $systemModuleUuid,
-        public string $entityId,
-        public mixed $data = null,
-    ) {
-    }
-
-    public static function viaRequest(array $data): self
-    {
-        return new self(
-            name: (string) $data['name'],
-            systemModuleUuid: (string) $data['system_module_uuid'],
-            entityId: (string) $data['entity_id'],
-            data: $data['data'],
-        );
-    }
 }
