+++ src/Kanvas/SystemModules/Contracts/SystemModuleInputInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kanvas\SystemModules\Contracts;
+
+interface SystemModuleInputInterface
+{
+}
