+++ src/Kanvas/SystemModules/Repositories/SystemModulesRepository.php
@@ -7,7 +7,12 @@ namespace Kanvas\SystemModules\Repositories;
 use Baka\Traits\SearchableTrait;
 use Illuminate\Database\Eloquent\Model;
 use Kanvas\Apps\Models\Apps;
+use Kanvas\Companies\Models\Companies;
+use Kanvas\Exceptions\InternalServerErrorException;
+use Kanvas\SystemModules\Contracts\SystemModuleInputInterface;
 use Kanvas\SystemModules\Models\SystemModules;
+use Kanvas\Users\Models\Users;
+use Kanvas\Users\Repositories\UsersRepository;
 
 class SystemModulesRepository
 {
@@ -22,8 +27,6 @@ class SystemModulesRepository
      * Get System Module by its model_name.
      *
      * @param string $model_name
-     *
-     * @return SystemModules
      */
     public static function getByModelName(string $modelName, ?Apps $app = null): SystemModules
     {
@@ -43,10 +46,6 @@ class SystemModulesRepository
 
     /**
      * Get by name.
-     *
-     * @param string $name
-     *
-     * @return SystemModules
      */
     public static function getByName(string $name, ?Apps $app = null): SystemModules
     {
@@ -55,5 +54,71 @@ class SystemModulesRepository
         return SystemModules::where('name', $name)
                                     ->where('apps_id', $app->getKey())
                                     ->firstOrFail();
+    }
+
+    /**
+     * Get the entity from the input
+     */
+    public static function getEntityFromInput(SystemModuleInputInterface $entityInput, Users $user): Model
+    {
+        $systemModule = SystemModules::where('uuid', $entityInput->systemModuleUuid)
+                        ->fromApp()
+                        ->notDeleted()
+                        ->firstOrFail();
+
+        /**
+        * @var BaseModel
+        */
+        $entityModel = (new ($systemModule->model_name));
+        $hasUuid = $entityModel->hasColumn('uuid');
+
+        $isUser = $entityModel instanceof Users;
+        $isCompany = $entityModel instanceof Companies;
+
+        $hasAppId = $entityModel->hasColumn('apps_id');
+        $hasCompanyId = $entityModel->hasColumn('companies_id');
+
+        if (! $hasAppId && ! $hasCompanyId && (! $isUser && ! $isCompany)) {
+            throw new InternalServerErrorException('This system module doesn\'t allow external custom fields');
+        }
+        $field = $hasUuid ? 'uuid' : 'id';
+
+        if ($isUser || $isCompany) {
+            $entity = $entityModel::where('uuid', $entityInput->entityId)
+                    ->notDeleted()
+                    ->firstOrFail();
+
+            if ($user->isAppOwner()) {
+                return $entity;
+            }
+
+            //check if the user belongs to the company
+            if ($entity instanceof Users) {
+                UsersRepository::belongsToCompany(
+                    $entity,
+                    $user->getCurrentCompany()
+                );
+            } elseif ($entity instanceof Companies) {
+                UsersRepository::belongsToCompany(
+                    $user,
+                    $entity
+                );
+            }
+        } else {
+            if ($user->isAppOwner()) {
+                $entity = $entityModel::where($field, $entityInput->entityId)
+                        ->fromApp()
+                        ->notDeleted()
+                        ->firstOrFail();
+            } else {
+                $entity = $entityModel::where($field, $entityInput->entityId)
+                        ->fromApp()
+                        ->fromCompany($user->getCurrentCompany())
+                        ->notDeleted()
+                        ->firstOrFail();
+            }
+        }
+
+        return $entity;
     }
 }
